+++ examples/data-sources/nxos_default_qos_policy_map/data-source.tf
@@ -1,0 +1,3 @@
+data "nxos_default_qos_policy_map" "example" {
+  name = "PM1"
+}