+++ examples/resources/nxos_default_qos_policy_map/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_default_qos_policy_map.example "sys/ipqos/dflt/p/name-[PM1]"