+++ examples/resources/nxos_default_qos_policy_map/resource.tf
@@ -1,0 +1,4 @@
+resource "nxos_default_qos_policy_map" "example" {
+  name       = "PM1"
+  match_type = "match-any"
+}