+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - Add nxos_pim_ssm_range resource and data source
 - Add nxos_default_qos_class_map resource and data source
 - Add nxos_default_qos_class_map_dscp resource and data source
+- Add nxos_default_qos_policy_map resource and data source
 
 ## 0.2.3